qgis_cadtools: circulararc.py meldete fehler wegen float statt int bei ptcount
git-svn-id: http://cataisrepository.googlecode.com/svn/trunk/qgis/plugins/cadtools@296 5cb8516b-641c-f86c-07e9-861e6a0574f2

+++ tools/circulararc.py
@@ -60,7 +60,7 @@ class CircularArc:
         else:
             sweep = 0.0;
 
-        ptcount = math.ceil( math.fabs ( sweep / arcIncr ) )
+        ptcount = int(math.ceil( math.fabs ( sweep / arcIncr ) ))
 
         if sweep < 0: 
             arcIncr *= -1.0;
